Port DOM and Diff
Use a combination of the Swift port and the original Elixir versions.  While `Element`s are mutable because of `Jsoup`, this library itself is immutable as much by using persistent collections and checking if a change actually occurred when a diff is deepMerged.  Unlike the Elixir or Swift ports, specific types are used for Diff and its nested elements instead of Maps with Any values.